+++ src/data/postgres/tables/profiles.test.ts
@@ -17,12 +17,12 @@ import {
 import { RPSError } from '../../../utils/RPSError.js'
 import { ProfilesTable } from './profiles.js'
 import { jest } from '@jest/globals'
-import { type SpyInstance, spyOn } from 'jest-mock'
+import { type Spied, spyOn } from 'jest-mock'
 
 describe('profiles tests', () => {
   let db: PostgresDb
   let profilesTable: ProfilesTable
-  let querySpy: SpyInstance<any>
+  let querySpy: Spied<any>
   let amtConfig: AMTConfiguration
   const profileName = 'profileName'
   const tenantId = 'tenantId'
@@ -109,8 +109,8 @@ describe('profiles tests', () => {
       querySpy.mockResolvedValueOnce({ rows, rowCount: rows.length })
       const result = await profilesTable.get()
       expect(result).toStrictEqual(rows)
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
     SELECT 
       p.profile_name as "profileName", 
@@ -170,8 +170,8 @@ describe('profiles tests', () => {
       querySpy.mockResolvedValueOnce({ rows: [{}], rowCount: rows.length })
       const result = await profilesTable.getByName(profileName)
       expect(result).toStrictEqual(rows[0])
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
     SELECT 
       p.profile_name as "profileName",
@@ -252,8 +252,8 @@ describe('profiles tests', () => {
       expect(result).toBeTruthy()
       expect(wirelessConfigSpy).toHaveBeenLastCalledWith(profileName, tenantId)
       expect(profileProxyConfigsSpy).toHaveBeenLastCalledWith(profileName, tenantId)
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
       DELETE 
       FROM profiles 
@@ -304,8 +304,8 @@ describe('profiles tests', () => {
       )
 
       expect(getByNameSpy).toHaveBeenCalledWith(amtConfig.profileName, amtConfig.tenantId)
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
         INSERT INTO profiles(
           profile_name, activation,
@@ -355,8 +355,8 @@ describe('profiles tests', () => {
         amtConfig.tenantId
       )
       expect(getByNameSpy).toHaveBeenCalledWith(amtConfig.profileName, amtConfig.tenantId)
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
         INSERT INTO profiles(
           profile_name, activation,
@@ -420,8 +420,8 @@ describe('profiles tests', () => {
 
       const result = await profilesTable.update(amtConfig)
       expect(result).toBe(amtConfig)
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
       UPDATE profiles 
       SET activation=$2, amt_password=$3, generate_random_password=$4, cira_config_name=$5,
@@ -466,8 +466,8 @@ describe('profiles tests', () => {
 
       const result = await profilesTable.update(amtConfig)
       expect(result).toBe(amtConfig)
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
       UPDATE profiles 
       SET activation=$2, amt_password=$3, generate_random_password=$4, cira_config_name=$5,
@@ -542,8 +542,8 @@ describe('profiles tests', () => {
       )
 
       expect(getByNameSpy).toHaveBeenCalledWith(amtConfig.profileName, amtConfig.tenantId)
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
         INSERT INTO profiles(
           profile_name, activation,
@@ -593,8 +593,8 @@ describe('profiles tests', () => {
         amtConfig.tenantId
       )
       expect(getByNameSpy).toHaveBeenCalledWith(amtConfig.profileName, amtConfig.tenantId)
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
         INSERT INTO profiles(
           profile_name, activation,
@@ -639,8 +639,8 @@ describe('profiles tests', () => {
 
       const result = await profilesTable.update(amtConfig)
       expect(result).toBe(amtConfig)
-      expect(querySpy).toBeCalledTimes(1)
-      expect(querySpy).toBeCalledWith(
+      expect(querySpy).toHaveBeenCalledTimes(1)
+      expect(querySpy).toHaveBeenCalledWith(
         `
       UPDATE profiles 
       SET activation=$2, amt_password=$3, generate_random_password=$4, cira_config_name=$5,